llm/requirements.txt
fastapi
uvicorn
pyyaml
openai
python-dotenv
rsqlalchemy
aiomysql
langchain-community
langchain
langchain-openai
pinecone
langchain-pinecone